Avoid casting results of Math.round
That's not needed as long as the input is a float.

+++ app/src/main/java/com/kamron/pogoiv/logic/IVCombination.java
@@ -17,7 +17,7 @@ public class IVCombination {
         this.att = att;
         this.def = def;
         this.sta = sta;
-        percentPerfect = (int) Math.round(((att + def + sta) / 45.0) * 100);
+        percentPerfect = Math.round((att + def + sta) / 45f * 100);
     }
 
     @Override
